ESC-823 rename cache package to persistence
