+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/persistence/EoriEmailRepository.scala
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.eusubsidycompliancefrontend.cache
+package uk.gov.hmrc.eusubsidycompliancefrontend.persistence
 
 import org.mongodb.scala.model.{Filters, FindOneAndUpdateOptions, ReturnDocument, Updates}
 import org.mongodb.scala.result.UpdateResult
-import uk.gov.hmrc.eusubsidycompliancefrontend.cache.EoriEmailRepository.DefaultTtl
+import uk.gov.hmrc.eusubsidycompliancefrontend.persistence.EoriEmailRepository.DefaultTtl
 import uk.gov.hmrc.eusubsidycompliancefrontend.models.VerifiedEmail
 import uk.gov.hmrc.eusubsidycompliancefrontend.models.types.EORI
 import uk.gov.hmrc.mongo.cache.{CacheItem, MongoCacheRepository}
