+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/persistence/RepositoryHelpers.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.eusubsidycompliancefrontend.cache
+package uk.gov.hmrc.eusubsidycompliancefrontend.persistence
 
 import uk.gov.hmrc.mongo.cache.DataKey
