+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/services/EmailVerificationService.scala
@@ -22,7 +22,7 @@ import play.api.http.Status.CREATED
 import play.api.mvc.Results.Redirect
 import play.api.mvc.{AnyContent, Call, Result}
 import uk.gov.hmrc.eusubsidycompliancefrontend.actions.requests.AuthenticatedEnrolledRequest
-import uk.gov.hmrc.eusubsidycompliancefrontend.cache.EoriEmailRepository
+import uk.gov.hmrc.eusubsidycompliancefrontend.persistence.EoriEmailRepository
 import uk.gov.hmrc.eusubsidycompliancefrontend.connectors.EmailVerificationConnector
 import uk.gov.hmrc.eusubsidycompliancefrontend.models._
 import uk.gov.hmrc.eusubsidycompliancefrontend.models.types.EORI
