+++ test/uk/gov/hmrc/eusubsidycompliancefrontend/services/EmailVerificationServiceSpec.scala
@@ -26,7 +26,7 @@ import play.api.mvc.{AnyContent, Call}
 import play.api.test.Helpers._
 import play.api.test.{DefaultAwaitTimeout, FakeRequest}
 import uk.gov.hmrc.eusubsidycompliancefrontend.actions.requests.AuthenticatedEnrolledRequest
-import uk.gov.hmrc.eusubsidycompliancefrontend.cache.EoriEmailRepository
+import uk.gov.hmrc.eusubsidycompliancefrontend.persistence.EoriEmailRepository
 import uk.gov.hmrc.eusubsidycompliancefrontend.connectors.EmailVerificationConnector
 import uk.gov.hmrc.eusubsidycompliancefrontend.models.{EmailVerificationRequest, VerifiedEmail}
 import uk.gov.hmrc.eusubsidycompliancefrontend.syntax.FutureSyntax.FutureOps
